Refactor packet handling with new delegate structure
Introduce `PacketHandlerDelegate<TPacket>` to replace
`PacketHandlerInvoker<TPacket>`, enhancing performance
through compiled delegates for zero-allocation execution.
Update method signatures and variable types accordingly.
Remove or comment out several middleware-related lines
to simplify the middleware pipeline, improving
efficiency and maintainability of packet handling.
