+++ src/Nalix.Network/Dispatch/Middleware/Core/PacketMiddlewareDelegate.cs
@@ -1,6 +1,6 @@
 ﻿using Nalix.Network.Dispatch.Core;
 
-namespace Nalix.Network.Dispatch.Middleware;
+namespace Nalix.Network.Dispatch.Middleware.Core;
 
 /// <summary>
 /// Delegate type representing a middleware handler for processing packets.
