+++ src/Nalix.Network/Dispatch/Middleware/PacketMiddlewarePipeline.cs
@@ -1,4 +1,5 @@
 ﻿using Nalix.Network.Dispatch.Core;
+using Nalix.Network.Dispatch.Middleware.Core;
 
 namespace Nalix.Network.Dispatch.Middleware;
